// src/mixins/Border/Border.types.ts
export interface IBorder {
  border?: string;
  borderBottom?: string;
  borderBottomColor?: string | 'transparent'; // TODO: map to theme colors
  borderBottomStyle?: string;
  borderBottomWidth?: string;
  borderColor?: string | 'transparent'; // TODO: map to theme colors
  borderLeft?: string;
  borderLeftColor?: string | 'transparent'; // TODO: map to theme colors
  borderLeftStyle?: string;
  borderLeftWidth?: string;
  borderRight?: string;
  borderRightColor?: string | 'transparent'; // TODO: map to theme colors
  borderRightStyle?: string;
  borderRightWidth?: string;
  borderStyle?: string;
  borderTop?: string;
  borderTopColor?: string | 'transparent'; // TODO: map to theme colors
  borderTopStyle?: string;
  borderTopWidth?: string;
  borderWidth?: string;
  borderX?: string;
  borderXColor?: string | 'transparent'; // TODO: map to theme colors
  borderXStyle?: string;
  borderXWidth?: string;
  borderY?: string;
  borderYColor?: string | 'transparent'; // TODO: map to theme colors
  borderYStyle?: string;
  borderYWidth?: string;
}